Complete record_search rename to follow naming convention
Main's rename left the tool's types half-converted: types/search.ts kept
its old filename and SearchResult/SearchEvent/SearchToolResponse kept
their un-prefixed names while only SearchInput became RecordSearchInput.

This finishes the convention: types/search.ts -> types/record-search.ts
(the file is record_search-specific — only record_search files import it),
and SearchResult/SearchEvent/SearchToolResponse -> RecordSearch*. Upstream
FamilySearch API types (FSSearchResponse, etc.) and the generic TreeMatch
helper stay as-is. Spec updated to match.
